+++ src/core/component/functional/index.ts
@@ -16,5 +16,4 @@ export * from 'core/component/functional/interface';
 
 //#if runtime has dummyComponents
 import('core/component/functional/test/b-functional-dummy');
-import('core/component/functional/test/b-functional-button-dummy');
 //#endif